Rework font_size to scale the cell's pixel size, not the denomination
font_size was modelled as changing the root cell denomination (units per
cell), so raw-unit dimensions did NOT track it and a cascade of demo-side
workarounds was needed. The correct model: a cell's PIXEL size is set by
font_size while the denomination stays the authored default (8x16 at the
root); a unit is 1/denomination of a cell, so every unit-measured length
auto-scales in pixels as the cell grows, and the layout is unchanged in
units.

Raster backend:
- pixels-per-unit becomes font_size-aware (scale * font_size/12) and, so
  cell-aligned geometry stays crisp, cell-snapped per axis: whole
  denomination-cells map to exact integer cell-pixel multiples, the
  sub-cell remainder interpolates fractionally (pxX/pxY). Proportional
  text glyphs and decorative lengths use the unsnapped fraction (pxLen).
- SetFontSize replaces the CellMetricsForFontSize-drives-denomination
  approach; Metrics() (the denomination) stays 8x16 under font_size.
- text.Render takes a fractional pixels-per-unit.

Core: Painter exposes font_size-aware device-pixel helpers (PxPerUnitF,
UnitsToPx, deviceAnchor) via the new UnitPixelMapper backend capability,
so the device-pixel escape hatches (FillRectPixels, DrawImagePixels,
resize-hover, menu hairlines) line up with painted geometry at any
font_size. MeasureRunes counts units again (8 per rune, 16 for Tuesday) -
a unit count is font_size-invariant.

Host: kittytk-sdl / SDL platform pass font_size to every surface's
backend via SetFontSize; the UI font stays one cell tall in units.

Demo: drop the denominate preprocessor and the denomination-aware window
sizes; fixedbox is a raw unit width again (height-for-width kept).

PurfecTerm's font_size behaviour is deferred (only the text.Render
signature widened); the terminal renders as before.

+++ text/engine_test.go
@@ -284,7 +284,7 @@ func TestRenderProducesInk(t *testing.T) {
 
 	for _, scale := range []int{1, 2} {
 		img := image.NewRGBA(image.Rect(0, 0, int(sp.Width())*scale+4, int(sp.Height())*scale+4))
-		Render(img, sp, 0, 0, scale, color.RGBA{255, 255, 255, 255})
+		Render(img, sp, 0, 0, float64(scale), color.RGBA{255, 255, 255, 255})
 
 		ink := 0
 		bounds := img.Bounds()
